requirements.txt
appdirs==1.4.3
asgi-redis==1.0.0
asgiref==1.1.1
asn1crypto==0.22.0
attrs==16.3.0
autobahn==0.18.1
Automat==0.5.0
Babel==2.4.0
bleach==2.0.0
cairocffi==0.8.0
CairoSVG==2.0.2
cffi==1.10.0
channels==1.1.3
click==6.7
constantly==15.1.0
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
cryptography==1.8.1
cssselect==1.0.1
cycler==0.10.0
daphne==1.2.0
decorator==4.0.11
Django==1.10.6
django-autocomplete-light==3.2.7
django-bootstrap3==8.2.2
django-crispy-forms==1.6.1
django-custom-user-migration==0.3.0
django-fm==0.2.5
django-markdown2==0.3.0
django-phonenumber-field==1.3.0
django-pygments==0.1
django-tinymce==2.6.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
easy-thumbnails==2.4.1
entrypoints==0.2.2
gunicorn==19.7.1
html5lib==0.999999999
idna==2.5
incremental==16.10.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itypes==1.1.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lxml==3.7.3
markdown2==2.3.3
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.7.4
msgpack-python==0.4.8
nbconvert==5.1.1
nbformat==4.3.0
nltk==3.2.2
notebook==5.0.0
numpy==1.12.1
olefile==0.44
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
parsel==1.1.0
pexpect==4.2.1
phonenumberslite==8.4.3
pickleshare==0.7.4
Pillow==4.1.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
Pygments==2.2.0
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.2.0
Pyphen==0.9.4
PyPrind==2.10.0
pyt==0.7.26
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
queuelib==1.4.2
redis==2.10.5
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
Scrapy==1.3.3
scrapy-djangoitem==1.1.1
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
terminado==0.6
testpath==0.3
tinycss==0.4
tornado==4.5
traitlets==4.3.2
Twisted==17.1.0
txaio==2.6.2
uritemplate==3.0.0
w3lib==1.17.0
wcwidth==0.1.7
WeasyPrint==0.36
webencodings==0.5.1
Werkzeug==0.12.1
whitenoise==3.3.0
widgetsnbextension==2.0.0
zope.interface==4.3.3
